+++ src/models/DefaultPlugin.js
@@ -93,6 +93,13 @@ class DefaultPlugin {
   }
 
   /**
+   * Reset actions.
+   */
+  resetDrawerActions() {
+    this.__drawer.resetDrawerActions();
+  }
+
+  /**
    * Draws all data in the html element defined by the id.
    * @param {string} id - Html id, without '#'.
    * @param {boolean} readOnly - Make the draw read-only.